pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "AssetPixelHand"
description = "An AI-powered tool for finding and managing duplicate and visually similar images."
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [{ name = "kilazz"}]

dependencies = [
    # Core Application & GUI
    "PySide6>=6.10.0",
    "send2trash>=1.8.3",
    "pywin32>=311; sys_platform == 'win32'",
    # AI & Machine Learning
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "onnx>=1.19.1",
    "sentencepiece>=0.2.1",
    "huggingface_hub[hf_xet]>=0.36.0", #transformers
    # Databases
    "duckdb>=1.4.1",
    "lancedb>=0.25.2",
    # Image & Data Handling
    "scipy>=1.16.3",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "numpy>=2.2.6", #oiio-python
    "xxhash>=3.6.0",
    "ImageHash>=4.3.2",
    "zstandard>=0.25.0",
    # Image Processing
    "Pillow>=12.0.0",
    #"OpenImageIO>=3.0.11.0",
    "oiio-python>=3.0.10.0.1", #replacement OpenImageIO
    #"directxtex-decoder", #For DDS
]

[project.optional-dependencies]
cpu = ["onnxruntime>=1.23.2"]
gpu = ["onnxruntime-directml>=1.23.0"]

dev = [
    "ruff",
    "snakeviz",
    "tabulate2",
    "types-oiio-python>=3.0.10.0.1",
]

[tool.ruff]
exclude = [
    ".bzr", ".direnv", ".eggs", ".git", ".git-rewrite", ".hg",
    ".ipynb_checkpoints", ".mypy_cache", ".nox", ".pants.d", ".pyenv",
    ".pytest_cache", ".pytype", ".ruff_cache", ".svn", ".tox", ".venv",
    ".vscode", "__pypackages__", "_build", "buck-out", "build", "dist",
    "node_modules", "site-packages", "venv", "app_data", "_src_directxtex_decoder", "_src_pillow-simd"
]

line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle: errors (e.g., indentation, syntax)
    "W",   # pycodestyle: warnings (e.g., line length)
    "F",   # Pyflakes: logical errors (undefined names, unused imports)
    "I",   # isort: sorting and formatting of imports
    "B",   # flake8-bugbear: finds likely bugs and design problems
    "UP",  # pyupgrade: upgrades syntax to newer Python versions
    "SIM", # flake8-simplify: suggests simplifications for complex code
    "RUF", # Enables Ruff's native, fast, and often unique checks
]

ignore = ["E501"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["app_data*"]
